wikitest: automatic tests on the wiki documentation
This procedure test wiki pages against master or release repositories.
Assuming that wiki.sympy and release version are placed near master
repository, e.g.:

    ./sympy/
    ./wiki.sympy/
    ./sympy-0.6.7/

If the directories are not the same, then they can be tuned by options.

The files are to be passing through the test contain the special directives:

* reStructedText:  .. wikitest release
* markdown:        <!-- wikitest release -->